pyproject.toml
[project]
name = "coldcall"
version = "0.1.0"
description = "Benchmark and test voice AI agents via cold calls"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11,<3.14"
authors = [
    { name = "Kollaikal Rupesh" },
]
keywords = ["voice", "ai", "testing", "benchmark", "twilio", "pipecat", "stt", "tts"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Communications :: Telephony",
]
dependencies = [
    "twilio>=9.0",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "websockets>=14.0",
    "pipecat-ai[websocket,deepgram,openai,cartesia,silero]>=0.0.105",
    "typer[all]>=0.9",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/kollaikal-rupesh/coldcall"
Repository = "https://github.com/kollaikal-rupesh/coldcall"
Issues = "https://github.com/kollaikal-rupesh/coldcall/issues"
Documentation = "https://github.com/kollaikal-rupesh/coldcall#readme"

[project.scripts]
coldcall = "coldcall.cli:main"

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["coldcall"]

[tool.hatch.build.targets.wheel.force-include]
"scenarios" = "coldcall/builtin_scenarios"
